requirements.txt
flask
opencv-python
numpy
Pillow
torch
torchvision
Flask-Bootstrap 